requirements.txt
psycopg2-binary
sqlalchemy
pandas
requests
pdfplumber
pdf2image
paddleocr
paddlepaddle
reportlab
python-dateutil
schedule
numpy
opencv-python
streamlit
plotly
google-auth
google-auth-oauthlib
google-api-python-client
openpyxl
